some debugging of posreg and timer

+++ node2/controllers/include/posController.h
@@ -1,6 +1,6 @@
 struct PID_data {
   int32_t error_sum;
-  float Kp;
+  int8_t Kp;
   int8_t Ki;
   int16_t position;
   float sample_time;
@@ -10,6 +10,6 @@ struct PID_data {
   int16_t encoder_value;
 };
 
-void pos_controller_init(float p_factor, int8_t i_factor, float sample_time, uint8_t encoder_factor);
+void pos_controller_init(int8_t p_factor, int8_t i_factor, float sample_time, uint8_t encoder_factor);
 void pos_controller_calculate_power(int8_t reference_value, int16_t measured_value);
 int16_t pos_controller_get_power();
